Made Changes to Animation Editors

+++ PizaroAnimator/Animation2D.cs
@@ -45,12 +45,12 @@ namespace Zeroit.Framework.Transitions.ZeroitPizaroAnimator
         /// <summary>
         /// Start point of the animation.
         /// </summary>
-        protected double[] m_Start;
+        protected float[] m_Start;
 
         /// <summary>
         /// End point of the animation.
         /// </summary>
-        protected double[] m_End;
+        protected float[] m_End;
 
         /// <summary>
         /// Duration of the animation in milliseconds.
@@ -61,7 +61,7 @@ namespace Zeroit.Framework.Transitions.ZeroitPizaroAnimator
         /// Acceleration function, which must return a number between 0 and 1 for inputs between 0
         /// and 1.
         /// </summary>
-        private Func<double, double> m_AccelFunc;
+        private Func<float, float> m_AccelFunc;
 
         /// <summary>
         /// Acceleration function, which must return a number between 0 and 1 for inputs between 0
@@ -73,7 +73,7 @@ namespace Zeroit.Framework.Transitions.ZeroitPizaroAnimator
         /// Acceleration function, which must return a number between 0 and 1 for inputs between 0
         /// and 1.
         /// </summary>
-        private double acceleration;
+        private float acceleration;
 
         /// <summary>
         /// Timestamp for when the animation was started.
@@ -98,7 +98,7 @@ namespace Zeroit.Framework.Transitions.ZeroitPizaroAnimator
         /// <summary>
         /// Current co-ordinates for animation.
         /// </summary>
-        protected double[] m_Current;
+        protected float[] m_Current;
 
         /// <summary>
         /// Current state of the animation.
@@ -108,7 +108,7 @@ namespace Zeroit.Framework.Transitions.ZeroitPizaroAnimator
         /// <summary>
         /// Percent of the way through the animation (between 0.0 and 1.0).
         /// </summary>
-        protected double m_Progress;
+        protected float m_Progress;
 
         private System.Timers.Timer m_Timer;
 
@@ -119,7 +119,7 @@ namespace Zeroit.Framework.Transitions.ZeroitPizaroAnimator
         /// <param name="end">Array of end co-ordinates.</param>
         /// <param name="duration">Length of animation in milliseconds.</param>
         /// <param name="accelFunc">Acceleration function, returns 0-1 for inputs 0-1.</param>
-        //public Animation2D(double[] start, double[] end, int duration, Func<double, double> accelFunc) {
+        //public Animation2D(float[] start, float[] end, int duration, Func<float, float> accelFunc) {
         //    if (start.Length != end.Length) {
         //        throw new AnimationException("Start and end arrays must be the same length.");
         //    }
@@ -128,14 +128,14 @@ namespace Zeroit.Framework.Transitions.ZeroitPizaroAnimator
         //    m_Duration = duration;
         //    m_AccelFunc = accelFunc;
 
-        //    m_Current = new double[m_Start.Length];
+        //    m_Current = new float[m_Start.Length];
         //    m_State = AnimationState.Stopped;
         //    m_Progress = 0;
         //    m_Timer = new System.Timers.Timer(TIMEOUT);
         //    m_Timer.Elapsed += new ElapsedEventHandler(m_Timer_Elapsed);
         //}
 
-        public Animation2D(double[] start, double[] end, int duration, double easing)
+        public Animation2D(float[] start, float[] end, int duration, float easing)
         {
             if (start.Length != end.Length)
             {
@@ -146,7 +146,7 @@ namespace Zeroit.Framework.Transitions.ZeroitPizaroAnimator
             m_Duration = duration;
             acceleration = easing;
 
-            m_Current = new double[m_Start.Length];
+            m_Current = new float[m_Start.Length];
             m_State = AnimationState.Stopped;
             m_Progress = 0;
             m_Timer = new System.Timers.Timer(TIMEOUT);
@@ -176,14 +176,14 @@ namespace Zeroit.Framework.Transitions.ZeroitPizaroAnimator
         /// <summary>
         /// Gets the current co-ordinates of the animation.
         /// </summary>
-        public double[] CurrentCoordinates {
+        public float[] CurrentCoordinates {
             get { return m_Current; }
         }
 
         /// <summary>
         /// Gets the progress of the animation (a number between 0.0 and 1.0).
         /// </summary>
-        public double Progress {
+        public float Progress {
             get { return m_Progress; }
         }
 
@@ -281,7 +281,7 @@ namespace Zeroit.Framework.Transitions.ZeroitPizaroAnimator
         /// </summary>
         /// <param name="now">The current time.</param>
         private void Cycle(DateTime now) {
-            m_Progress = (now - m_StartTime).TotalMilliseconds / (m_EndTime - m_StartTime).TotalMilliseconds;
+            m_Progress = (float)(now - m_StartTime).TotalMilliseconds / (float)(m_EndTime - m_StartTime).TotalMilliseconds;
 
             if (m_Progress >= 1) {
                 m_Progress = 1;
@@ -322,11 +322,11 @@ namespace Zeroit.Framework.Transitions.ZeroitPizaroAnimator
         /// Calculates current co-ordinates, based on the current state.
         /// </summary>
         /// <param name="t">Percentage of the way through the animation as a decimal.</param>
-        private void UpdateCoords(double t) {
+        private void UpdateCoords(float t) {
             if (t == 1) {
                 m_Current = m_End;
             } else {
-                m_Current = new double[m_Start.Length];
+                m_Current = new float[m_Start.Length];
                 for (int i = 0; i < m_Start.Length; i++) {
                     m_Current[i] = (m_End[i] - m_Start[i]) * t + m_Start[i];
                 }
